pyproject.toml
[project]
name = "electro-chain"
version = "0.1.0"
description = ""
authors = [{ name = "Nikolay Korkushko, email = <kakashonochka12@gmail.com>" }]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "drf-yasg (>=1.21.11,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
black = "^25.9.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"
isort = "^7.0.0"


[tool.poetry.group.testing.dependencies]
coverage = "^7.11.0"

